requirements.txt
asgiref==3.5.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.48.0
django-summernote==0.8.20.0
django-taggit==2.1.0
idna==3.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
